sched/fair: reserve prime core for high_cap tasks only
There are cases when random tasks from either root or other non top-app
cpusets end up running on prime core even though there is no reason for
them to do so. They are just waking the damn core for no reason.

Let's try and fix that byt sending them to mid cores, from which
they can be sent to little (or stay on mid if they need to).

Signed-off-by: Yaroslav Furman <yaro330@gmail.com>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>
Signed-off-by: Carlos Ayrton Lopez Arroyo <15030201@itcelaya.edu.mx>